+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor.hs
@@ -454,7 +454,7 @@ peerSelectionGovernor :: ( Alternative (STM m)
                       -> m Void
 peerSelectionGovernor tracer debugTracer countersTracer fuzzRng stateVar actions policy =
     JobPool.withJobPool $ \jobPool -> do
-      localPeers <- map (\(target, _) -> (target, 0))
+      localPeers <- map (\(h, w, _) -> (h, w, 0))
                 <$> atomically (readLocalRootPeers actions)
       peerSelectionGovernorLoop
         tracer